requirements.txt
streamlit>=1.28.1
openai>=1.3.7
python-dotenv>=1.0.0
reportlab>=4.0.7
Pillow>=10.1.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-openai>=0.0.2
google-generativeai